+++ enterprise/internal/codeintel/autoindexing/transport/graphql/utils.go
@@ -2,7 +2,6 @@ package graphql
 
 import (
 	"encoding/base64"
-	"fmt"
 	"strconv"
 	"strings"
 
@@ -11,10 +10,8 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend/graphqlutil"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/autoindexing/shared"
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/shared/types"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	resolverstubs "github.com/sourcegraph/sourcegraph/internal/codeintel/resolvers"
-	"github.com/sourcegraph/sourcegraph/lib/codeintel/autoindex/config"
 )
 
 // strPtr creates a pointer to the given value. If the value is an
@@ -25,42 +22,6 @@ func strPtr(val string) *string {
 	}
 
 	return &val
-}
-
-// getKeyForLookup creates a quick unique key for a map lookup.
-func getKeyForLookup(indexer, root string) string {
-	return fmt.Sprintf("%s:%s", indexer, root)
-}
-
-type availableIndexer struct {
-	Roots []string
-	URL   string
-}
-
-type JobsOrHints interface {
-	config.IndexJob | config.IndexJobHint
-	GetIndexerName() string
-	GetRoot() string
-}
-
-func populateInferredAvailableIndexers[J JobsOrHints](jobsOrHints []J, blocklist map[string]struct{}, inferredAvailableIndexers map[string]availableIndexer) map[string]availableIndexer {
-	for _, job := range jobsOrHints {
-		indexer := job.GetIndexerName()
-		key := getKeyForLookup(indexer, job.GetRoot())
-		// Only add them to the inferred jobs map if they're not already in the recent uploads
-		// blocklist. This is to avoid hinting at an available index if we've already indexed it.
-		if _, ok := blocklist[key]; !ok {
-			ai := inferredAvailableIndexers[key]
-			ai.Roots = append(ai.Roots, job.GetRoot())
-			if p, ok := types.PreferredIndexers[indexer]; ok {
-				ai.URL = fmt.Sprintf("https://%s", p.URN)
-			}
-
-			inferredAvailableIndexers[indexer] = ai
-		}
-	}
-
-	return inferredAvailableIndexers
 }
 
 func marshalLSIFIndexGQLID(indexID int64) graphql.ID {
